Story 2 and 3
Added the checksum as well as the executable test program ParserTest.exe
and a release version Parser.exe. Please see the README file for how to
use the release version. Story 3 is complete as well when specifying an
input and output file from the command line for the release version.