Refactored AccountDigit class and fixed checksum
Re-factored the AccountDigit struct into a class with its own files.
Reduced a lot of code usage this way and tests were good enough to catch
any mistakes.
Modified checksum tests for additional values and Fixed an off by one
found in the checksum.
Added some print statements around the possible error codes for a file
that is not in the right format to be parsed in the release
configuration.